Add assignment operator to TestFoo and TestBar test util classes.
I added a an assignment operator to the classes, and I added a field
named `magic` to each of them that gets set to certain value during
construction. Only an object that has been constructed is likely to
have that magic value properly set.

An assignment precondition is that the left-hand side is valid.
The assignment operators I added assert that the `magic` member of
the lhs has the right value.